+++ apps/mobile/android/app/src/main/java/io/github/c1oudreamw/lumatile/MainActivity.kt
@@ -1,4 +1,4 @@
-package cn.edu.qlu.toolbox
+package io.github.c1oudreamw.lumatile
 
 import android.os.Bundle
 import android.view.WindowManager
@@ -8,6 +8,7 @@ class MainActivity : BridgeActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         window.addFlags(WindowManager.LayoutParams.FLAG_SECURE)
         registerPlugin(GradeExportPlugin::class.java)
+        registerPlugin(AppUpdatePlugin::class.java)
         super.onCreate(savedInstanceState)
     }
 }